+++ src/index.js
@@ -2,14 +2,13 @@ import ReactDOM from "react-dom/client";
 import { StrictMode } from "react";
 import "./index.scss";
 import App from "./components/App/App";
-import ThemeContextProvider from "./context/ThemeContext";
+import MainContextProvider from "./context/MainContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <StrictMode>
-    <ThemeContextProvider>
+    <MainContextProvider>
       <App />
-      <p>test husky3</p>
-    </ThemeContextProvider>
+    </MainContextProvider>
   </StrictMode>,
 );
