--- src/utils/constants.js
@@ -1,6 +1,0 @@
-export const THEMECOLOR = {
-  LIGHT: "light",
-  DARK: "dark",
-};
-
-export const TEST = 0;
